Stubbed in some Extern work to simply listen for towerInstruction.
Easiest external case to handle: the extern device sends towerInstruction
packets, indicating they want the Towers to do some stuff.  The Console
notes this, sets a 10 second timer to prevent gamePlay from happending,
and plays the towerInstruction.lightLevel as button presses on the Console.
fireLevel is ignored, but whatever.  The Towers, of course, do play out
the fireLevel.