Console.ino updated to listen to remote relays (Fire and System enable/disable).
Wand.ino got a minor change to use the right radio library.  Major change is
in Console.ino, where I stub out "listening" for Fire enable/disable
remote control.  On a state change, Console makes an audible alert for
the Operator.  I also stubbed out "listening" for System enable/disable
remote control.  The idea is for the Console to start up in "test/setup"
mode, where we can jigger with settings and test the system components.
When the relay is tripped, the Console shifts to "normal/gameplay" mode.